MAINT: remove code duplication
Move some experiment-independent plotting code to
experiments/plotting.py
