+++ chemex/experiments/cpmg/plotting.py
@@ -8,38 +8,7 @@ from matplotlib import ticker
 from matplotlib.backends import backend_pdf
 import numpy as np
 
-from chemex import peaks
 from chemex.experiments import plotting
-
-
-def sigma_estimator(x):
-    """Estimates standard deviation using median to exclude outliers.
-
-    Up to 50% can be bad.
-
-    """
-    return np.median([np.median(abs(xi - np.asarray(x))) for xi in x]) * 1.1926
-
-
-def set_lim(values, scale):
-    """Provide a range that contains all the value and adds a margin."""
-    v_min, v_max = min(values), max(values)
-    margin = (v_max - v_min) * scale
-    v_min, v_max = v_min - margin, v_max + margin
-
-    return v_min, v_max
-
-
-def group_data(dataset):
-    """Group the data resonance specifically."""
-    data_grouped = dict()
-
-    for profile in dataset:
-        resonance_id = profile.profile_name
-        peak = peaks.Peak(resonance_id)
-        data_grouped[peak] = profile
-
-    return data_grouped
 
 
 def compute_profiles(data_grouped, params):
@@ -109,9 +78,9 @@ def plot_data(data, params, output_dir='./'):
 
         print(("  * {} [.fit]".format(name_pdf)))
 
-        data_grouped = group_data(dataset)
+        data_grouped = plotting.group_data(dataset)
         profiles, r2_min, r2_max = compute_profiles(data_grouped, params)
-        ymin, ymax = set_lim([r2_min, r2_max], 0.10)
+        ymin, ymax = plotting.set_lim([r2_min, r2_max], 0.10)
 
         with backend_pdf.PdfPages(name_pdf) as file_pdf, open(name_txt, 'w') as file_txt:
 
@@ -147,7 +116,7 @@ def plot_data(data, params, output_dir='./'):
                     markerfacecolor='None',
                     zorder=3, )
 
-                xmin, xmax = set_lim(nu_cpmg, 0.10)
+                xmin, xmax = plotting.set_lim(nu_cpmg, 0.10)
 
                 ax2.set_xlim(xmin, xmax)
                 ax2.set_ylim(ymin, ymax)
@@ -177,8 +146,8 @@ def plot_data(data, params, output_dir='./'):
                     markerfacecolor='None',
                     zorder=100, )
 
-                rmin, _rmax = set_lim(deltas - r2_erd, 0.2)
-                _rmin, rmax = set_lim(deltas + r2_eru, 0.2)
+                rmin, _rmax = plotting.set_lim(deltas - r2_erd, 0.2)
+                _rmin, rmax = plotting.set_lim(deltas + r2_eru, 0.2)
 
                 ax1.set_xlim(xmin, xmax)
                 ax1.set_ylim(rmin, rmax)
